Shuffle things around to resolve compile warnings
prtest:full
