+++ crates/wasmtime/src/runtime/vm/gc/enabled.rs
@@ -7,13 +7,12 @@ mod externref;
 mod free_list;
 mod structref;
 #[cfg(any(feature = "gc-drc", feature = "gc-copying"))]
-mod trace_info;
+mod trace_infos;
 
 pub use arrayref::*;
 pub use exnref::*;
 pub use externref::*;
 pub use structref::*;
-pub(crate) use trace_info::TraceInfo;
 
 #[cfg(feature = "gc-drc")]
 mod drc;
